Delete volumes online without restarting volume server
