+++ weed/storage/store.go
@@ -338,3 +338,13 @@ func (s *Store) UnmountVolume(i VolumeId) error {
 	return fmt.Errorf("Volume %d not found on disk", i)
 }
 
+func (s *Store) DeleteVolume(i VolumeId) error {
+	for _, location := range s.Locations {
+		if error := location.deleteVolumeById(i); error == nil {
+			s.updateMaster()
+			return nil
+		}
+	}
+
+	return fmt.Errorf("Volume %d not found on disk", i)
+}
